+++ Examples/User_Parser/User_Parser.ino
@@ -173,7 +173,7 @@ void loop()
 
 // Call back from within parser, for end of message
 // Process a complete message incoming from parser
-void userMessage(SEMP_PARSE_STATE *parse, uint8_t type)
+void userMessage(SEMP_PARSE_STATE *parse, uint16_t type)
 {
     USER_SCRATCH_PAD *scratchPad = (USER_SCRATCH_PAD *)parse->scratchPad;
     static bool displayOnce = true;
